AD-796: Adding Python SDK that can help customers communicate with our services (#508)
AD-796: Adding Python SDK that can help customers communicate with our services

AD-796: Improving the manager that handles message timeouts.

- Using a thread Event to synchronize the thread and wait for the timeout to occur which is way cleaner than a while loop.


NOTE: Since this is currently just part of our monorepo, we don't yet have a way (in this PR) to distribute this to customers effectively. A follow-up PR will add the scripts to package this up and publish it to PyPi for consumption.